% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corte.R
\name{corte}
\alias{corte}
\title{Title}
\usage{
corte(bd, variable, corte, cuantil = FALSE, paleta)
}
\arguments{
\item{bd}{tibble. Es la tabla a la que le agregaremos la columna 'color'}

\item{variable}{nse. Es la variable en formato 'non standard evaluation' a la que
se va a cortar por cuantil}

\item{corte}{vector int. Vector de cortes base para hacer el corte de colores}

\item{cuantil}{lgl. Indica si el vector corte debará traducirse en cuantiles de
'variable'}

\item{paleta}{vector chr. Vector de colores. La cantidad de colores depende
de la cantidad de la longitud del parámetro 'corte' (length(corte)-1)}
}
\value{

}
\description{
Title
}
\examples{
Ejemplo
bd <- tibble(a = rnorm(50))
bd \%>\% cuantil(a, cuantil = c(0,.5,1), paleta = c("red","green"))


}
