% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradado.R
\name{rampa}
\alias{rampa}
\title{Title}
\usage{
rampa(bd, variable, color_min, color_max)
}
\arguments{
\item{color_max}{}
}
\value{

}
\description{
Title
}
\examples{
Ejemplo
library(dplyr)
library(ggplot2)
library(grDevices)

bd <- tibble(a = letters, b= rnorm(length(letters)))

bd \%>\% rampa(b, "white","red") \%>\%
ggplot(aes(x = b, y = fct_reorder(a,b), fill = color)) + geom_col( ) +
scale_fill_identity()
}
